requirements.txt
pytest>=5.3.5
requests>=2.21
pillow>=7.0
jsonschema>=3.0.1
beautifulsoup4>=4.8.2
tinydb>=3.15.2
pydantic
flask>=2.0.0
streamlit>=1.40.0
ddddocr
numpy
